🩹 Buttons check, trimming
Hide contents that are not ready, updated buttons in forms to use Router
